fix(review): address code-review findings (wave 1)
- AI popup Retry now replays the actual request (was a no-op for errors)
- Cancel in-flight OpenAI request when the reader is dismissed
- Screenshot crop: untransformed thumbnail orientation (consistent w/ full-res path)
- Crop file resolver: symlink canonicalization parity with the scheme handler
- Mokuro import: uniquify dir on title collision (no longer wipes a different book);
  dedupe true same-volume re-import
- EPUB import now stamps contentType/importedAt (sync forward-compat)
- HTTP sync list: encode query params like Android URLEncoder (/ and : percent-encoded)
- payload manifest sizeBytes Int64; single security-scope owner on manga import
- friendlier manga page label when page count unknown

+++ Features/HttpSync/SyncBlobs.swift
@@ -165,7 +165,7 @@ nonisolated struct HttpSyncAiChatSettingsBlob: Codable, Equatable {
 /// Manifest describing a book's payload.zip.
 nonisolated struct HttpSyncPayloadManifest: Codable, Equatable {
     let sha256: String
-    let sizeBytes: Int
+    let sizeBytes: Int64
     let originalName: String
     let format: ContentType
 }
